+++ daemon/oci_windows.go
@@ -89,15 +89,10 @@ func (daemon *Daemon) createSpec(c *container.Container) (*libcontainerd.Spec, e
 
 	// s.Windows.LayerPaths
 	var layerPaths []string
-	if img.RootFS != nil && (img.RootFS.Type == image.TypeLayers || img.RootFS.Type == image.TypeLayersWithBase) {
+	if img.RootFS != nil && img.RootFS.Type == image.TypeLayers {
 		// Get the layer path for each layer.
-		start := 1
-		if img.RootFS.Type == image.TypeLayersWithBase {
-			// Include an empty slice to get the base layer ID.
-			start = 0
-		}
 		max := len(img.RootFS.DiffIDs)
-		for i := start; i <= max; i++ {
+		for i := 1; i <= max; i++ {
 			img.RootFS.DiffIDs = img.RootFS.DiffIDs[:i]
 			path, err := layer.GetLayerPath(daemon.layerStore, img.RootFS.ChainID())
 			if err != nil {
