+++ distribution/pull_v2_windows.go
@@ -3,37 +3,14 @@
 package distribution
 
 import (
-	"encoding/json"
-	"fmt"
 	"net/http"
 	"os"
 
 	"github.com/docker/distribution"
 	"github.com/docker/distribution/context"
-	"github.com/docker/distribution/manifest/schema1"
 	"github.com/docker/distribution/manifest/schema2"
 	"github.com/docker/distribution/registry/client/transport"
-	"github.com/docker/docker/image"
 )
-
-func detectBaseLayer(is image.Store, m *schema1.Manifest, rootFS *image.RootFS) error {
-	v1img := &image.V1Image{}
-	if err := json.Unmarshal([]byte(m.History[len(m.History)-1].V1Compatibility), v1img); err != nil {
-		return err
-	}
-	if v1img.Parent == "" {
-		return fmt.Errorf("Last layer %q does not have a base layer reference", v1img.ID)
-	}
-	// There must be an image that already references the baselayer.
-	for _, img := range is.Map() {
-		if img.RootFS.Type == image.TypeLayersWithBase && img.RootFS.BaseLayerID() == v1img.Parent {
-			rootFS.BaseLayer = img.RootFS.BaseLayer
-			rootFS.Type = image.TypeLayersWithBase
-			return nil
-		}
-	}
-	return fmt.Errorf("Invalid base layer %q", v1img.Parent)
-}
 
 var _ distribution.Describable = &v2LayerDescriptor{}
